# BUG: spatial/distance: fix pdist/cdist performance regression with w=None