Specifying w=None should be the same as not specifying weights.
Fix the code to respect this, and not select the very much slower
pure-Python code path for w=None.

Fixes the 5000x regressions... https://pv.github.io/scipy-bench/#regressions?sort=3&dir=desc

asv (for one --- running all of them is too slow because the pure-Python code path takes minutes for some inputs, when the C version is done in a few milliseconds):
```
$ python3 runtests.py --bench-compare=master -- 'spatial.Xdist.time_.*cityblock.*'
...
       before           after         ratio
     [10233db2]       [561eb57c]
     <master>         <distance-fix-w>
-       754±100μs       13.7±0.3μs     0.02  spatial.Xdist.time_pdist(10, 'cityblock')
-      1.42±0.2ms         20.9±6μs     0.01  spatial.Xdist.time_cdist(10, 'cityblock')
-        59.9±3ms       30.1±0.3μs     0.00  spatial.Xdist.time_pdist(100, 'cityblock')
-        141±30ms         52.9±9μs     0.00  spatial.Xdist.time_cdist(100, 'cityblock')
-      12.3±0.01s      3.27±0.06ms     0.00  spatial.Xdist.time_cdist(1000, 'cityblock')
```