@charris 2018-09-06
Note that the pytest 3.8.0 was just released, might be related.