@perimosocordiae 2018-09-08
Merged. Should we consider a backport?