@pv 2018-09-08
At the moment it doesn't look likely there will be a `1.1.1` release, and since this only occurs with `w=None` but not when omitting `w` from kwargs, it's probably not high enough priority.